Projeto Jira Python/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
rich==13.7.1
# Ollama roda localmente — sem dependência extra
